requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pytz==2024.2
PyYAML==6.0.2
simplejwt==2.0.1
sqlparse==0.5.2
typing==3.7.4.3
tzdata==2024.2
uritemplate==4.1.1
